+++ src/collective/volto/gdprcookie/restapi/deserializer.py
@@ -1,16 +1,16 @@
-from collective.volto.gdprcookie.interfaces import IGDPRCookieSettings
+from collective.volto.gdprcookie.interfaces import IGDPRCookieSettingsControlpanel
+from collective.volto.gdprcookie.restapi import parse_gdpr_blocks
 from plone.restapi.deserializer import json_body
 from plone.restapi.deserializer.controlpanels import ControlpanelDeserializeFromJson
+from plone.restapi.interfaces import IBlockFieldDeserializationTransformer
 from plone.restapi.interfaces import IDeserializeFromJson
 from zExceptions import BadRequest
 from zope.component import adapter
 from zope.interface import implementer
 
-import json
-
 
 @implementer(IDeserializeFromJson)
-@adapter(IGDPRCookieSettings)
+@adapter(IGDPRCookieSettingsControlpanel)
 class GDPRCookieSettingsDeserializeFromJson(ControlpanelDeserializeFromJson):
     def __call__(self):
         """
@@ -29,8 +29,13 @@ class GDPRCookieSettingsDeserializeFromJson(ControlpanelDeserializeFromJson):
                 }
             )
             raise BadRequest(errors)
+        gdpr_cookie_settings = parse_gdpr_blocks(
+            context=self.context,
+            data=gdpr_cookie_settings,
+            transformer_interface=IBlockFieldDeserializationTransformer,
+        )
         try:
-            setattr(proxy, "gdpr_cookie_settings", json.dumps(gdpr_cookie_settings))
+            setattr(proxy, "gdpr_cookie_settings", gdpr_cookie_settings)
         except ValueError as e:
             errors.append(
                 {