+++ src/collective/volto/gdprcookie/restapi/get.py
@@ -1,5 +1,7 @@
 from collective.volto.gdprcookie.interfaces import IGDPRCookieSettings
+from collective.volto.gdprcookie.restapi import parse_gdpr_blocks
 from plone import api
+from plone.restapi.interfaces import IBlockFieldSerializationTransformer
 from plone.restapi.interfaces import IExpandableElement
 from plone.restapi.services import Service
 from zope.component import adapter
@@ -41,6 +43,15 @@ class GDPRCookieSettings:
             # product not installed
             return None
 
+    def get_gdpr_cookies_value(self):
+        gdpr_cookies_settings = self.get_field_value(field="gdpr_cookie_settings")
+        data = parse_gdpr_blocks(
+            context=self.context,
+            data=gdpr_cookies_settings,
+            transformer_interface=IBlockFieldSerializationTransformer,
+        )
+        return json.loads(data)
+
     def get_data(self):
         """
         Return settings
@@ -50,10 +61,7 @@ class GDPRCookieSettings:
             return {}
         show_icon = self.get_field_value(field="show_icon")
         technical_cookies_only = self.get_field_value(field="technical_cookies_only")
-        gdpr_cookie_settings = self.get_field_value(field="gdpr_cookie_settings")
-        if gdpr_cookie_settings:
-            gdpr_cookie_settings = json.loads(gdpr_cookie_settings)
-
+        gdpr_cookie_settings = self.get_gdpr_cookies_value()
         data = {
             "show_icon": show_icon,
             "cookie_version": self.get_field_value(field="cookie_version"),