+++ src/collective/volto/gdprcookie/restapi/serializer.py
@@ -1,4 +1,7 @@
-from collective.volto.gdprcookie.interfaces import IGDPRCookieSettings
+from collective.volto.gdprcookie.interfaces import IGDPRCookieSettingsControlpanel
+from collective.volto.gdprcookie.restapi import parse_gdpr_blocks
+from plone import api
+from plone.restapi.interfaces import IBlockFieldSerializationTransformer
 from plone.restapi.interfaces import ISerializeToJson
 from plone.restapi.serializer.controlpanels import ControlpanelSerializeToJson
 from zope.component import adapter
@@ -12,15 +15,14 @@ logger = logging.getLogger(__name__)
 
 
 @implementer(ISerializeToJson)
-@adapter(IGDPRCookieSettings)
+@adapter(IGDPRCookieSettingsControlpanel)
 class GDPRCookieSettingsSerializeToJson(ControlpanelSerializeToJson):
     def __call__(self):
         json_data = super().__call__()
         gdpr_cookie_settings = json_data["data"].get("gdpr_cookie_settings", "{}")
-        try:
-            json_data["data"]["gdpr_cookie_settings"] = json.loads(gdpr_cookie_settings)
-        except json.decoder.JSONDecodeError:
-            logger.error(
-                f"Unable to convert value into json object: {gdpr_cookie_settings}"
-            )
+        json_data["data"]["gdpr_cookie_settings"] = parse_gdpr_blocks(
+            context=api.portal.get(),
+            data=gdpr_cookie_settings,
+            transformer_interface=IBlockFieldSerializationTransformer,
+        )
         return json_data