+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix serializer/deserializer to use block transformers.
+  [cekk]
 
 
 1.0.0 (2024-02-13)